Move the uniswap lp token balance querying code to the app
It should eventually be moved out of there and into the uniswap module.
