+++ tools/uniswapv2/detect.py
@@ -2,11 +2,10 @@ import argparse
 import json
 import os
 from pathlib import Path
-from typing import Any, Dict, List, NamedTuple, Optional
+from typing import Any, Dict, Optional
 
 from eth_utils.address import to_checksum_address
 
-from rotkehlchen.accounting.structures import Balance
 from rotkehlchen.chain.ethereum.contracts import EthereumContract
 from rotkehlchen.chain.ethereum.graph import Graph
 from rotkehlchen.chain.ethereum.manager import (
@@ -14,14 +13,11 @@ from rotkehlchen.chain.ethereum.manager import (
     EthereumManager,
     NodeName,
 )
-from rotkehlchen.chain.ethereum.utils import multicall_specific, token_normalized_value_decimals
-from rotkehlchen.chain.ethereum.zerion import DefiBalance, Zerion
-from rotkehlchen.constants.ethereum import ZERION_ABI
-from rotkehlchen.constants.misc import ZERO
+from rotkehlchen.chain.ethereum.utils import multicall_specific
+from rotkehlchen.chain.ethereum.zerion import Zerion, uniswap_lp_token_balances
 from rotkehlchen.externalapis.etherscan import Etherscan
 from rotkehlchen.greenlets import GreenletManager
 from rotkehlchen.tests.utils.ethereum import wait_until_all_nodes_connected
-from rotkehlchen.typing import ChecksumEthAddress
 from rotkehlchen.user_messages import MessagesAggregator
 from rotkehlchen.utils.misc import get_chunks, ts_now
 
@@ -71,68 +67,6 @@ def pairs_from_ethereum(ethereum: EthereumManager) -> Dict[str, Any]:
         pairs.extend([to_checksum_address(x[0]) for x in result])
 
     return pairs
-
-
-class UniLPBalance(NamedTuple):
-    pool: DefiBalance
-    token0: DefiBalance
-    token1: DefiBalance
-
-
-def _decode_token(entry) -> DefiBalance:
-    decimals = entry[0][3]
-    return DefiBalance(
-        token_address=entry[0][0],
-        token_name=entry[0][1],
-        token_symbol=entry[0][2],
-        balance=Balance(
-            amount=token_normalized_value_decimals(entry[1], decimals),
-            usd_value=ZERO,
-        ),
-    )
-
-
-def _decode_result(data) -> UniLPBalance:
-    pool_token = _decode_token(data[0])
-    token0 = _decode_token(data[1][0])
-    token1 = _decode_token(data[1][1])
-    return UniLPBalance(pool=pool_token, token0=token0, token1=token1)
-
-
-def token_balances_from_ethereum(
-        address: ChecksumEthAddress,
-        ethereum: EthereumManager,
-        zerion: Zerion,
-        lp_addresses: List[ChecksumEthAddress],
-):
-    """Query uniswap token balances from ethereum
-
-    The number of addresses to query in one call depends a lot on the node used.
-    With an infura node we saw the following:
-    500 addresses per call took on average 43 seconds for 20450 addresses
-    2000 addresses per call took on average 36 seconds for 20450 addresses
-    4000 addresses per call took on average 32.6 seconds for 20450 addresses
-    5000 addresses timed out a few times
-    """
-    zerion_contract = EthereumContract(
-        address=zerion.contract_address,
-        abi=ZERION_ABI,
-        deployed_block=0,
-    )
-    chunks = list(get_chunks(lp_addresses, n=4000))
-    balances = []
-    for idx, chunk in enumerate(chunks):
-        print(f'Querying univ2 token balances for {address} {idx + 1} / {len(chunks)}')
-        result = zerion_contract.call(
-            ethereum=ethereum,
-            method_name='getAdapterBalance',
-            arguments=[address, '0x4EdBac5c8cb92878DD3fd165e43bBb8472f34c3f', chunk]
-        )
-
-        for entry in result[1]:
-            balances.append(_decode_result(entry))
-
-    return balances
 
 
 def pairs_and_token_details_from_graph() -> Dict[str, Any]:
@@ -262,8 +196,8 @@ if __name__ == "__main__":
             write_result_to_file(result, 'uniswap_lp_tokens_ethereum.json')
 
         start = ts_now()
-        balances = token_balances_from_ethereum(
-            address='0x1554d34D46842778999cB4eb1381b19f651e4a9d',
+        balances = uniswap_lp_token_balances(
+            address='0x1554d34D46842778999cB4eb1381b19f651e4a9d',  # test address
             ethereum=ethereum,
             zerion=zerion,
             lp_addresses=results['ethereum'],
